Switch to Polybar from Lemonbar
A few reasons for this:
1. lemonbar-xft has not been touched in many years, I've run into some
bugs (I forget which) which have been fixed upstream but not ported to
the fork
2. The font loading didn't seem to be working properly.  I was not
getting the expected fonts at all.
3. Polybar is efficient!  I wrote a custom Pulse monitor
(https://github.com/SevereOverfl0w/pulse-volume/) and found that Polybar
does it the exact same way.
4. Some of my custom Lemonbar scripts were wonky.  Occasionally my bspwm
integration would be missing a workspace or not show something it
should.  This seemed simpler.